move pil_to_base64 to utils
